miner/analytics/requirements.txt
requests
python-dotenv
pytz
tzlocal
ujson==5.*